requirements.txt
aiohttp>=3.9.0
python-telegram-bot>=21.0
requests>=2.31.0
gunicorn>=21.0.0
